[rename] 클래스 이름 간단화 (#29)
